# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Hekayaty World Universe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters/{id}:
    get:
      operationId: getCharactersId
      tags: [characters]
      summary: Get character detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharacterDetail"
        "404":
          description: Not Found

  # =========================================================================
  # Comics
  # =========================================================================
...
  /worlds/{id}:
    get:
      operationId: getWorldsId
      tags: [worlds]
      summary: Get world detail with regions and locations
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorldDetail"
        "404":
          description: Not Found

  # =========================================================================
  # Timeline
  # =========================================================================
...
  /encyclopedia/{id}:
    get:
      operationId: getEncyclopediaId
      tags: [encyclopedia]
      summary: Get encyclopedia entry detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncyclopediaEntryDetail"
        "404":
          description: Not Found

  # =========================================================================
  # News
  # =========================================================================
...
  /gallery:
    get:
      operationId: getGallery
      tags: [gallery]
      summary: List all media
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
  /upload:
    post:
      operationId: postUpload
      tags: [gallery]
      summary: Upload an image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  description: Base64 string of the image
                category:
                  type: string
                altText:
                  type: string
                folder:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
components:
  schemas:
...
    Media:
      type: object
      properties:
        id: { type: string }
        publicId: { type: string }
        secureUrl: { type: string }
        width: { type: number, nullable: true }
        height: { type: number, nullable: true }
        format: { type: string, nullable: true }
        resourceType: { type: string, nullable: true }
        category: { type: string }
        altText: { type: string, nullable: true }

    # ========================== Characters ==========================
    CharacterBase:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        arabicName: { type: string }
        alias: { type: string, nullable: true }
        title: { type: string, nullable: true }
        alignment: { type: string }
        status: { type: string }
        powerCategory: { type: string, nullable: true }
        shortBio: { type: string, nullable: true }
        portraitUrl: { type: string, nullable: true }
        portraitAlt: { type: string, nullable: true }
    CharacterStats:
      type: object
      properties:
        strength: { type: number }
        speed: { type: number }
        intelligence: { type: number }
        wisdom: { type: number }
        willpower: { type: number }
        magic: { type: number }
    CharacterAbility:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
    CharacterRelationship:
      type: object
      properties:
        id: { type: string }
        relationType: { type: string }
        description: { type: string, nullable: true }
        relatedId: { type: string }
        relatedName: { type: string }
        relatedArabicName: { type: string }
        relatedPortraitUrl: { type: string, nullable: true }
    CharacterDetail:
      allOf:
        - $ref: "#/components/schemas/CharacterBase"
        - type: object
          properties:
            quote: { type: string, nullable: true }
            organization: { type: string, nullable: true }
            fullBio: { type: string, nullable: true }
            aboutText: { type: string, nullable: true }
            stats: { $ref: "#/components/schemas/CharacterStats", nullable: true }
            abilities: { type: array, items: { $ref: "#/components/schemas/CharacterAbility" } }
            relationships: { type: array, items: { $ref: "#/components/schemas/CharacterRelationship" } }
            gallery: { type: array, items: { $ref: "#/components/schemas/Media" } }
            appearances:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  role: { type: string, nullable: true }
                  issueId: { type: string }
                  issueTitle: { type: string }
                  issueArabicTitle: { type: string, nullable: true }
                  issueNumber: { type: number }
                  seriesTitle: { type: string }
                  coverUrl: { type: string, nullable: true }
            worlds:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  relationship: { type: string, nullable: true }
                  worldId: { type: string }
                  worldName: { type: string }
                  worldArabicName: { type: string }
                  coverUrl: { type: string, nullable: true }
            artifacts:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  entryId: { type: string }
                  title: { type: string }
                  arabicTitle: { type: string, nullable: true }
                  category: { type: string }
                  coverUrl: { type: string, nullable: true }
            timeline:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  eventId: { type: string }
                  yearLabel: { type: string }
                  title: { type: string }
                  arabicTitle: { type: string, nullable: true }
                  subtitle: { type: string, nullable: true }
                  eraName: { type: string, nullable: true }
                  eraArabicName: { type: string, nullable: true }

    # ========================== Comics ==========================
...
    WorldBase:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        arabicName: { type: string }
        description: { type: string, nullable: true }
        coverUrl: { type: string, nullable: true }
    WorldDetail:
      allOf:
        - $ref: "#/components/schemas/WorldBase"
        - type: object
          properties:
            regions:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  name: { type: string }
                  arabicName: { type: string, nullable: true }
                  locations:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        arabicName: { type: string, nullable: true }
                        locationType: { type: string, nullable: true }
            characters:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  relationship: { type: string, nullable: true }
                  characterId: { type: string }
                  characterName: { type: string }
                  characterArabicName: { type: string }
                  portraitUrl: { type: string, nullable: true }
            gallery:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  secureUrl: { type: string }
                  altText: { type: string, nullable: true }

    # ========================== Timeline ==========================
...
    EncyclopediaEntryBase:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        arabicTitle: { type: string, nullable: true }
        category: { type: string }
        coverUrl: { type: string, nullable: true }
    EncyclopediaEntryDetail:
      allOf:
        - $ref: "#/components/schemas/EncyclopediaEntryBase"
        - type: object
          properties:
            content: { type: string, nullable: true }
            tags:
              type: array
              items:
                type: string
            relatedCharacters:
              type: array
              items:
                type: object
                properties:
                  characterId: { type: string }
                  name: { type: string }
                  arabicName: { type: string }
                  portraitUrl: { type: string, nullable: true }
            relatedWorlds:
              type: array
              items:
                type: object
                properties:
                  worldId: { type: string }
                  name: { type: string }
                  arabicName: { type: string }

    # ========================== News ==========================
...